refactor(e2e): explode multi-describe consensus suites
- Explode consensus/block_building.parallel.test.ts (2 describes) into:
    consensus/block_building/simple.test.ts
    consensus/block_building/high_tps.test.ts
  with shared setup in consensus/block_building/setup.ts.
- Explode consensus/mbps.parallel.test.ts (7 describes) into:
    consensus/mbps/proposed_anchor.test.ts
    consensus/mbps/l2_to_l1.test.ts
    consensus/mbps/l1_to_l2.test.ts
    consensus/mbps/non_validator_sync.test.ts
    consensus/mbps/deploy_and_call.test.ts
    consensus/mbps/pipelining.test.ts
    consensus/mbps/redistribution.parallel.test.ts  (2 its — keeps .parallel)
  with shared helpers in consensus/mbps/setup.ts.
- Update bootstrap.sh discovery to add consensus/block_building/*.test.ts
  and consensus/mbps/*.test.ts globs.
- Re-point .test_patterns.yml redistribution entry to new path.
